Move RequiresDatabase and RequiresComponents to api package
