+++ src/main/kotlin/io/github/freya022/botcommands/api/components/Buttons.kt
@@ -1,5 +1,6 @@
 package io.github.freya022.botcommands.api.components
 
+import io.github.freya022.botcommands.api.components.annotations.RequiresComponents
 import io.github.freya022.botcommands.api.components.builder.button.ButtonFactory
 import io.github.freya022.botcommands.api.components.utils.ButtonContent
 import io.github.freya022.botcommands.api.core.service.annotations.BService
