+++ src/main/kotlin/io/github/freya022/botcommands/api/pagination/Paginators.kt
@@ -2,7 +2,7 @@ package io.github.freya022.botcommands.api.pagination
 
 import io.github.freya022.botcommands.api.components.Button
 import io.github.freya022.botcommands.api.components.Components
-import io.github.freya022.botcommands.api.components.RequiresComponents
+import io.github.freya022.botcommands.api.components.annotations.RequiresComponents
 import io.github.freya022.botcommands.api.core.BContext
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.Dependencies
