+++ src/main/kotlin/io/github/freya022/botcommands/internal/components/timeout/ComponentTimeoutHandlers.kt
@@ -1,8 +1,8 @@
 package io.github.freya022.botcommands.internal.components.timeout
 
 import io.github.freya022.botcommands.api.components.Components
-import io.github.freya022.botcommands.api.components.RequiresComponents
 import io.github.freya022.botcommands.api.components.annotations.ComponentTimeoutHandler
+import io.github.freya022.botcommands.api.components.annotations.RequiresComponents
 import io.github.freya022.botcommands.api.components.data.ComponentTimeoutData
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.Dependencies
