+++ src/main/kotlin/io/github/freya022/botcommands/internal/core/db/InternalDatabase.kt
@@ -1,6 +1,7 @@
 package io.github.freya022.botcommands.internal.core.db
 
 import io.github.freya022.botcommands.api.core.db.Database
+import io.github.freya022.botcommands.api.core.db.annotations.RequiresDatabase
 import io.github.freya022.botcommands.api.core.service.annotations.BService
 import io.github.freya022.botcommands.api.core.service.annotations.Dependencies
 import io.github.freya022.botcommands.api.core.service.annotations.IgnoreServiceTypes
