+++ src/test/java/doc/java/examples/commands/slash/SlashSayAgainPersistent.java
@@ -8,8 +8,8 @@ import io.github.freya022.botcommands.api.commands.application.slash.annotations
 import io.github.freya022.botcommands.api.commands.application.slash.annotations.SlashOption;
 import io.github.freya022.botcommands.api.commands.application.slash.annotations.TopLevelSlashCommandData;
 import io.github.freya022.botcommands.api.components.Buttons;
-import io.github.freya022.botcommands.api.components.RequiresComponents;
 import io.github.freya022.botcommands.api.components.annotations.JDAButtonListener;
+import io.github.freya022.botcommands.api.components.annotations.RequiresComponents;
 import io.github.freya022.botcommands.api.components.event.ButtonEvent;
 import io.github.freya022.botcommands.test.switches.TestLanguage;
 import net.dv8tion.jda.api.interactions.components.buttons.Button;
