+++ src/test/kotlin/doc/kotlin/examples/commands/slash/SlashSelectRoleEphemeral.kt
@@ -6,8 +6,8 @@ import io.github.freya022.botcommands.api.commands.application.ApplicationComman
 import io.github.freya022.botcommands.api.commands.application.slash.GuildSlashEvent
 import io.github.freya022.botcommands.api.commands.application.slash.annotations.JDASlashCommand
 import io.github.freya022.botcommands.api.components.EntitySelectMenu
-import io.github.freya022.botcommands.api.components.RequiresComponents
 import io.github.freya022.botcommands.api.components.SelectMenus
+import io.github.freya022.botcommands.api.components.annotations.RequiresComponents
 import io.github.freya022.botcommands.api.core.utils.after
 import io.github.freya022.botcommands.test.switches.TestLanguage
 import net.dv8tion.jda.api.entities.Role
