+++ src/test/kotlin/io/github/freya022/botcommands/test/commands/slash/SlashDb.kt
@@ -7,12 +7,12 @@ import io.github.freya022.botcommands.api.commands.application.ApplicationComman
 import io.github.freya022.botcommands.api.commands.application.slash.GuildSlashEvent
 import io.github.freya022.botcommands.api.commands.application.slash.annotations.JDASlashCommand
 import io.github.freya022.botcommands.api.core.db.Database
+import io.github.freya022.botcommands.api.core.db.annotations.RequiresDatabase
 import io.github.freya022.botcommands.api.core.db.preparedStatement
 import io.github.freya022.botcommands.api.core.db.transactional
 import io.github.freya022.botcommands.api.core.db.withLogger
 import io.github.freya022.botcommands.api.core.service.annotations.Dependencies
 import io.github.freya022.botcommands.api.core.utils.loggerOf
-import io.github.freya022.botcommands.internal.core.db.RequiresDatabase
 import io.github.oshai.kotlinlogging.KotlinLogging
 
 @Command
