+++ src/test/kotlin/io/github/freya022/botcommands/test/services/ConditionalServiceTest.kt
@@ -1,7 +1,7 @@
 package io.github.freya022.botcommands.test.services
 
 import io.github.freya022.botcommands.api.components.Components
-import io.github.freya022.botcommands.api.components.RequiresComponents
+import io.github.freya022.botcommands.api.components.annotations.RequiresComponents
 import io.github.freya022.botcommands.api.core.service.ConditionalServiceChecker
 import io.github.freya022.botcommands.api.core.service.ServiceContainer
 import io.github.freya022.botcommands.api.core.service.annotations.BService
